+++ src/main/java/com/andersmmg/cityessentials/client/screen/EnvelopeScreen.java
@@ -1,0 +1,34 @@
+package com.andersmmg.cityessentials.client.screen;
+
+import com.andersmmg.cityessentials.CityEssentials;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.minecraft.client.gui.DrawContext;
+import net.minecraft.client.gui.screen.ingame.HandledScreen;
+import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.text.Text;
+import net.minecraft.util.Identifier;
+
+@Environment(EnvType.CLIENT)
+public class EnvelopeScreen extends HandledScreen<EnvelopeScreenHandler> {
+    private static final Identifier TEXTURE = CityEssentials.id("textures/gui/container/envelope.png");
+
+    public EnvelopeScreen(EnvelopeScreenHandler handler, PlayerInventory inventory, Text title) {
+        super(handler, inventory, title);
+        this.backgroundHeight = 133;
+//        this.playerInventoryTitleY = this.backgroundHeight - 94;
+        this.playerInventoryTitleY = 90000;
+    }
+
+    public void render(DrawContext context, int mouseX, int mouseY, float delta) {
+        this.renderBackground(context);
+        super.render(context, mouseX, mouseY, delta);
+        this.drawMouseoverTooltip(context, mouseX, mouseY);
+    }
+
+    protected void drawBackground(DrawContext context, float delta, int mouseX, int mouseY) {
+        int i = (this.width - this.backgroundWidth) / 2;
+        int j = (this.height - this.backgroundHeight) / 2;
+        context.drawTexture(TEXTURE, i, j, 0, 0, this.backgroundWidth, this.backgroundHeight);
+    }
+}