+++ src/main/java/com/andersmmg/cityessentials/datagen/ModModelProvider.java
@@ -25,6 +25,7 @@ public class ModModelProvider extends FabricModelProvider {
     public void generateItemModels(ItemModelGenerator itemModelGenerator) {
         itemModelGenerator.register(ModItems.PAPER_BAG, Models.GENERATED);
         itemModelGenerator.register(ModItems.GROCERY_BAG, Models.GENERATED);
+        itemModelGenerator.register(ModItems.ENVELOPE, Models.GENERATED);
 
         itemModelGenerator.register(ModItems.CASH_1, Models.GENERATED);
         itemModelGenerator.register(ModItems.CASH_5, Models.GENERATED);