+++ src/main/java/com/andersmmg/cityessentials/item/custom/BagItem.java
@@ -6,17 +6,23 @@ import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.screen.NamedScreenHandlerFactory;
 import net.minecraft.screen.SimpleNamedScreenHandlerFactory;
+import net.minecraft.sound.SoundCategory;
+import net.minecraft.sound.SoundEvent;
 import net.minecraft.util.Hand;
 import net.minecraft.util.TypedActionResult;
 import net.minecraft.world.World;
 
 public class BagItem extends Item {
-    public BagItem(Settings settings) {
+    public final SoundEvent openSound;
+
+    public BagItem(Settings settings, SoundEvent openSound) {
         super(settings);
+        this.openSound = openSound;
     }
 
     @Override
     public TypedActionResult<ItemStack> use(World world, PlayerEntity user, Hand hand) {
+        world.playSound(user, user.getBlockPos(), openSound, SoundCategory.PLAYERS, 1.0F, 1.0F);
         ItemStack stack = user.getStackInHand(hand);
         user.openHandledScreen(createScreenHandlerFactory(stack));
         return TypedActionResult.success(stack);