+++ src/main/java/com/andersmmg/cityessentials/sounds/ModSounds.java
@@ -13,6 +13,7 @@ public class ModSounds {
     public static final SoundEvent GROCERY_BAG_OPEN = registerSoundEvent("grocery_bag_open");
     public static final SoundEvent MAILBOX_OPEN = registerSoundEvent("mailbox_open");
     public static final SoundEvent MAILBOX_CLOSE = registerSoundEvent("mailbox_close");
+    public static final SoundEvent ENVELOPE_OPEN = registerSoundEvent("envelope_open");
 
     private static SoundEvent registerSoundEvent(String name) {
         Identifier id = CityEssentials.id(name);